Add files via upload
I added shapefiles for the city of Kuopio, primary and secondary education points, streets, and cycling and pedestrian paths, as well as a forest raster file for the municipality that Kuopio is in. I am still looking for the air quality measurement and a few more files for various facilities as well as the housing data.